Remove Id field from clientResponse.
Since it was of type uint64 it caused a panic when trying to decode
responses where the id was returned as a string.

The Id field is not used for anything and the clientResponse type in not
exported so there shouldn't be any negative consequences of this
removal.

See https://groups.google.com/d/topic/gorilla-web/JrkB7PfU-58/discussion