v2/json2: Support request params as object and array.
RPC 2.0 Specification allows for both request params
as object and as an array. Handle both cases.

Exceprts from JSON RPC 2.0 spec:

```
If present, parameters for the rpc call MUST be provided
as a Structured value. Either by-position through an Array
or by-name through an Object.

by-position:
params MUST be an Array, containing the values in the
Server expected order.

by-name:
params MUST be an Object, with member names that match the
Server expected parameter names. The absence of expected
names MAY result in an error being generated. The names
MUST match exactly, including case, to the method's
expected parameters.
```